Refactor: 1. Add https support 2. Modify part of api
